Make mock senso api request into the normal request.
