+++ src/main/java/controllers/CarDataController.java
@@ -6,6 +6,7 @@ import interfaces.ApiInputAdapter;
 import entities.Car;
 import org.springframework.web.bind.annotation.*;
 import usecases.CarToJsonRequestAdapter;
+import usecases.SensoReadyInfo;
 
 import java.io.IOException;
 import java.net.URI;
@@ -59,7 +60,7 @@ public class CarDataController {
 
     @PutMapping("/userCarLoan")
     public String userCarLoan(@RequestBody String inputData) throws IOException, InterruptedException {
-        return userCarLoanRequest(inputData);
+        return SensoReadyInfo.userCarLoanRequest(inputData);
     }
 
 
@@ -70,65 +71,7 @@ public class CarDataController {
         return car.getDepreciation();
     }
 
-//    /**
-//     * Turn a car into something that the frontend can recieve and work with.
-//     * @param car is the car you want to turn into a JSON.
-//     * @return a JSON representation of the car.
-//     */
-//    @NotNull
-//    private String makeCarJSON(Car car) {
-//        return getJsonFormattedCar(car);
-//    }
 
-    public static String userCarLoanRequest(String inputData) throws IOException, InterruptedException {
-
-        ObjectMapper mapper = new ObjectMapper();
-
-        try {
-            // convert JSON string to Map
-            Map<String, String> userInputs = mapper.readValue(inputData, Map.class);
-            System.out.println(userInputs);
-
-            double loanAmount = Double.parseDouble(userInputs.get("loanAmount"));
-            double pytBudget = Double.parseDouble(userInputs.get("pytBudget"));
-            int carId = Integer.parseInt(userInputs.get("carId"));
-            double downPayment = Double.parseDouble(userInputs.get("downPayment"));
-
-
-            String creditInputJson = ApiInputAdapter.makeCreditInputJSON();
-
-            var creditRequest = HttpRequest.newBuilder()
-                    .uri(URI.create("MCOK URI"))
-                    .header("Content-Type", "application/json")
-                    .header("x-api-key", "MOCK KEY")
-                    .POST(HttpRequest.BodyPublishers.ofString(creditInputJson))
-                    .build();
-            CreditScoreController csc = new CreditScoreController();
-            int creditScore = csc.pingCreditScoreAPI(creditRequest);
-
-            String loanInputJson = ApiInputAdapter.makeLoanInputJSON(loanAmount, creditScore, pytBudget, carId, downPayment);
-
-            var loanRequest = HttpRequest.newBuilder()
-                    .uri(URI.create(System.getenv("SENSO_URL")))
-                    .header("Content-Type", "application/json")
-                    .header("x-api-key", System.getenv("SENSO_KEY"))
-                    .POST(HttpRequest.BodyPublishers.ofString(loanInputJson))
-                    .build();
-
-            var client = HttpClient.newHttpClient();
-
-            var response = client.send(loanRequest, HttpResponse.BodyHandlers.ofString());
-
-            System.out.println(response.statusCode());
-            System.out.println(response.body());
-
-            return response.body();
-
-        } catch (IOException e) {
-            e.printStackTrace();
-            return "Server Error. Car Loan request";
-        }
-    }
 
 
     @GetMapping("/test")
