+++ src/main/java/controllers/CreditScoreController.java
@@ -1,10 +1,17 @@
 package controllers;
 
+import java.net.URI;
 import java.net.http.HttpRequest;
 
 public class CreditScoreController {
 
-    public int pingCreditScoreAPI(HttpRequest creditRequest){
+    public int pingCreditScoreAPI(String creditInputJson){
+        var creditRequest = HttpRequest.newBuilder()
+                .uri(URI.create("MCOK URI"))
+                .header("Content-Type", "application/json")
+                .header("x-api-key", "MOCK KEY")
+                .POST(HttpRequest.BodyPublishers.ofString(creditInputJson))
+                .build();
         return generateCreditScore();
     }
